e2e: move cluster retrieval functions to util/placement.go
- Moved functions responsible for retrieving the current cluster from dractions/retry.go to util/placement.go
- Moved getPlacement from dractions/crug.go to util/placement.go
- Refactored getTargetCluster to get drPolicy and also changed to pass currentCluster as argument.
- Made GetCurrentCluster(added doc) and GetPlacement functions public.
- Updated references to use the new public functions.
