+++ e2e/dractions/actions.go
@@ -41,16 +41,15 @@ func EnableProtection(ctx types.Context) error {
 	placementName := name
 	drpcName := name
 
-	placementDecision, err := waitPlacementDecision(util.Ctx.Hub.Client, managementNamespace, placementName)
+	clusterName, err := util.GetCurrentCluster(util.Ctx.Hub.Client, managementNamespace, placementName)
 	if err != nil {
 		return err
 	}
 
-	clusterName := placementDecision.Status.Decisions[0].ClusterName
 	log.Debugf("Workload running on cluster %q", clusterName)
 
 	err = retry.RetryOnConflict(retry.DefaultBackoff, func() error {
-		placement, err := getPlacement(util.Ctx.Hub.Client, managementNamespace, placementName)
+		placement, err := util.GetPlacement(util.Ctx.Hub.Client, managementNamespace, placementName)
 		if err != nil {
 			return err
 		}
@@ -117,21 +116,14 @@ func Failover(ctx types.Context) error {
 	log := ctx.Logger()
 	name := ctx.Name()
 
-	drpcName := name
 	client := util.Ctx.Hub.Client
-	drPolicyName := util.DefaultDRPolicyName
 
-	currentCluster, err := getCurrentCluster(client, managementNamespace, name)
+	currentCluster, err := util.GetCurrentCluster(client, managementNamespace, name)
 	if err != nil {
 		return err
 	}
 
-	drpolicy, err := util.GetDRPolicy(client, drPolicyName)
-	if err != nil {
-		return err
-	}
-
-	targetCluster, err := getTargetCluster(client, managementNamespace, drpcName, drpolicy)
+	targetCluster, err := getTargetCluster(client, currentCluster)
 	if err != nil {
 		return err
 	}
@@ -150,21 +142,14 @@ func Relocate(ctx types.Context) error {
 	log := ctx.Logger()
 	name := ctx.Name()
 
-	drpcName := name
 	client := util.Ctx.Hub.Client
-	drPolicyName := util.DefaultDRPolicyName
 
-	currentCluster, err := getCurrentCluster(client, managementNamespace, name)
+	currentCluster, err := util.GetCurrentCluster(client, managementNamespace, name)
 	if err != nil {
 		return err
 	}
 
-	drpolicy, err := util.GetDRPolicy(client, drPolicyName)
-	if err != nil {
-		return err
-	}
-
-	targetCluster, err := getTargetCluster(client, managementNamespace, drpcName, drpolicy)
+	targetCluster, err := getTargetCluster(client, currentCluster)
 	if err != nil {
 		return err
 	}
