+++ e2e/dractions/retry.go
@@ -4,7 +4,6 @@
 package dractions
 
 import (
-	"context"
 	"fmt"
 	"time"
 
@@ -14,37 +13,8 @@ import (
 	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"open-cluster-management.io/api/cluster/v1beta1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
-
-// waitPlacementDecision waits until we have a placement decision and returns the placement decision object.
-func waitPlacementDecision(client client.Client, namespace string, placementName string,
-) (*v1beta1.PlacementDecision, error) {
-	startTime := time.Now()
-
-	for {
-		placement, err := getPlacement(client, namespace, placementName)
-		if err != nil {
-			return nil, err
-		}
-
-		placementDecision, err := getPlacementDecisionFromPlacement(client, placement)
-		if err != nil {
-			return nil, err
-		}
-
-		if placementDecision != nil && len(placementDecision.Status.Decisions) > 0 {
-			return placementDecision, nil
-		}
-
-		if time.Since(startTime) > util.Timeout {
-			return nil, fmt.Errorf("timeout waiting for placement decisions for %q ", placementName)
-		}
-
-		time.Sleep(util.RetryInterval)
-	}
-}
 
 func waitDRPCReady(ctx types.Context, client client.Client, namespace string, drpcName string) error {
 	log := ctx.Logger()
@@ -109,15 +79,6 @@ func waitDRPCPhase(ctx types.Context, client client.Client, namespace, name stri
 	}
 }
 
-func getCurrentCluster(client client.Client, namespace string, placementName string) (string, error) {
-	placementDecision, err := waitPlacementDecision(client, namespace, placementName)
-	if err != nil {
-		return "", err
-	}
-
-	return placementDecision.Status.Decisions[0].ClusterName, nil
-}
-
 // return dr cluster client
 func getDRClusterClient(clusterName string, drpolicy *ramen.DRPolicy) client.Client {
 	if clusterName == drpolicy.Spec.DRClusters[0] {
@@ -127,8 +88,8 @@ func getDRClusterClient(clusterName string, drpolicy *ramen.DRPolicy) client.Cli
 	return util.Ctx.C2.Client
 }
 
-func getTargetCluster(client client.Client, namespace, placementName string, drpolicy *ramen.DRPolicy) (string, error) {
-	currentCluster, err := getCurrentCluster(client, namespace, placementName)
+func getTargetCluster(client client.Client, currentCluster string) (string, error) {
+	drpolicy, err := util.GetDRPolicy(client, util.DefaultDRPolicyName)
 	if err != nil {
 		return "", err
 	}
@@ -201,44 +162,4 @@ func waitDRPCProgression(
 
 		time.Sleep(util.RetryInterval)
 	}
-}
-
-func getPlacementDecisionFromPlacement(ctrlClient client.Client, placement *v1beta1.Placement,
-) (*v1beta1.PlacementDecision, error) {
-	matchLabels := map[string]string{
-		v1beta1.PlacementLabel: placement.GetName(),
-	}
-
-	listOptions := []client.ListOption{
-		client.InNamespace(placement.GetNamespace()),
-		client.MatchingLabels(matchLabels),
-	}
-
-	plDecisions := &v1beta1.PlacementDecisionList{}
-	if err := ctrlClient.List(context.Background(), plDecisions, listOptions...); err != nil {
-		return nil, fmt.Errorf("failed to list PlacementDecisions (placement: %s)",
-			placement.GetNamespace()+"/"+placement.GetName())
-	}
-
-	if len(plDecisions.Items) == 0 {
-		return nil, nil
-	}
-
-	if len(plDecisions.Items) > 1 {
-		return nil, fmt.Errorf("multiple PlacementDecisions found for Placement (count: %d, placement: %s)",
-			len(plDecisions.Items), placement.GetNamespace()+"/"+placement.GetName())
-	}
-
-	plDecision := plDecisions.Items[0]
-	// r.Log.Info("Found ClusterDecision", "ClsDedicision", plDecision.Status.Decisions)
-
-	if len(plDecision.Status.Decisions) > 1 {
-		return nil, fmt.Errorf("multiple placements found in PlacementDecision"+
-			" (count: %d, Placement: %s, PlacementDecision: %s)",
-			len(plDecision.Status.Decisions),
-			placement.GetNamespace()+"/"+placement.GetName(),
-			plDecision.GetName()+"/"+plDecision.GetNamespace())
-	}
-
-	return &plDecision, nil
 }
